+++ pkg/agent/common/secret_test.go
@@ -33,7 +33,7 @@ func TestGenerateSecretData(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := GenerateSecretData(tt.secretEnv)
+			got, err := GenerateSecretData(tt.secretEnv, "")
 			if (err != nil) != tt.wantErr {
 				t.Errorf("GenerateSecretData() error = %v, wantErr %v", err, tt.wantErr)
 				return