+++ pkg/grafana/service.go
@@ -1,24 +1,18 @@
 package grafana
 
 import (
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
 	"github.com/giantswarm/observability-operator/pkg/config"
 	grafanaClient "github.com/giantswarm/observability-operator/pkg/grafana/client"
 )
 
 type Service struct {
-	client        client.Client
 	grafanaClient grafanaClient.GrafanaClient
 	cfg           config.Config
 }
 
-func NewService(k8sClient client.Client, grafanaClient grafanaClient.GrafanaClient, cfg config.Config) *Service {
-	s := &Service{
-		client:        k8sClient,
+func NewService(grafanaClient grafanaClient.GrafanaClient, cfg config.Config) *Service {
+	return &Service{
 		grafanaClient: grafanaClient,
 		cfg:           cfg,
 	}
-
-	return s
 }